eink/requirements.txt
pillow>=8.4.0
RPi.GPIO>=0.7.0
spidev>=3.5